working on compile jsps
