+++ src/main/groovy/net/wasdev/wlp/gradle/plugins/tasks/AbstractServerTask.groovy
@@ -17,6 +17,7 @@ package net.wasdev.wlp.gradle.plugins.tasks
 
 import net.wasdev.wlp.gradle.plugins.extensions.DeployExtension
 import net.wasdev.wlp.gradle.plugins.extensions.LibertyExtension
+import org.gradle.api.GradleException
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
 import java.nio.file.Files
@@ -27,6 +28,8 @@ abstract class AbstractServerTask extends AbstractTask {
     private final String HEADER = "# Generated by liberty-gradle-plugin"
 
     def server
+
+    def libertyMap = buildLibertyMap(project)
 
     protected void executeServerCommand(Project project, String command, Map<String, String> params) {
         project.ant.taskdef(name: 'server',
@@ -242,4 +245,16 @@ abstract class AbstractServerTask extends AbstractTask {
             parentDir.mkdirs()
         }
     }
+
+    private String getPackagingType() throws Exception{
+      if (project.plugins.hasPlugin("war") || !project.tasks.withType(War).isEmpty()) {
+          return "war"
+      }
+      else if (project.plugins.hasPlugin("ear") || !project.tasks.withType(Ear).isEmpty()) {
+          return "ear"
+      }
+      else {
+          throw new GradleException("Archive path not found. Supported formats are jar, war, and ear.")
+      }
+  }
 }
